test(chore): unit tests for extractGraphQLAPIs
combine graphql schema extraction integration test cases so as to not re-run extraction multiple times
drop integration from test name
